Fix new highscore leaderboard bug.
When a new high score was a achieved it would
replace one of the old score entries. Now the
new score will be inserted between old entries
and the lower entries will be scooted down. The
last entry will be popped in order to maintain
the desired length of the leaderboard.