Include a check to make sure no UserId is saved.
A user's userId should not be saved into the table.
Only first names should be saved. A cleanupLeaderboards
method was added to be run after the user leaves the
skill to delete any entry with this UserId instead of
a name. Small changes were also made to leaderboard.js
to support this bug fix.